Increased format speed by ANOTHER 100% with ANOTHER one-line change.
Recall that LilyPondFile wraps around a ScoreBlock (just the way that ScoreBlock wraps around a Score).

So the exact same problem was occurring here:

    @property
    def _formatted_blocks(self):
        result = []
        for x in self:
            if 'lilypond_format' in dir(x) and not isinstance(x, str):
                lilypond_format = x.lilypond_format # <== this formats the score once
                if lilypond_format:
                    result.append(x.lilypond_format) # <== and this formats the score a second time!!
            else:
                result.append(str(x))
        return result


Which is changed in this commit with a single line:

    @property
    def _formatted_blocks(self):
        result = []
        for x in self:
            if 'lilypond_format' in dir(x) and not isinstance(x, str):
                lilypond_format = x.lilypond_format
                if lilypond_format:
                    result.append(lilypond_format) # <== this avoids reformatting the entire score
            else:
                result.append(str(x))
        return result


The increase in performance is yet another 100%.

The same segment of Geibete previously consumed not 10.2 million function calls but 20.4 
million function calls. (I was reading output from the wrong test script). 
That same segment now consumes only 5.1 million function calls:
ONLY A FOURTH OF NUMBER REQUIRED FOR THE PAST SEVERAL YEARS.

Unfuckingbelievable.