Set Expression.__hash__ equal to hash(repr(expression)).
Equal-valued expressions that haven't yet been evaluated always compare equal.

With this commit equal-valued expressions that haven't yet been evaluated now
return the same hash key, too.

The goal is to allow the interpreter to create a global hash of expressions
structured with unevaluated expressions as keys and the corresponding evaluated
expressions as values. This will then allow the interpreter to evaluate
an expression on first encounter and then bypass evaluation on all subsequent encounters.