Optimized sequencetools.repeat_sequence_to_weight_exactly().
Background: calling __add__ on NonreducedFraction is expensive.
This will need to be optimized in itself. But this commit contributes
a preliminary step in the form of a reimplementation of the (short)
function sequencetools.repeat_sequence_to_weight_exactly().
The old implementation of the function called mathtools.weight()
on the entire result sequence on each pass through a loop.
The new implementation of the function calls mathtools.weight() 
only once to calculate the number of total repetitions
of the sequence required. This avoids a large number of calls
to __add__ and therefore makes it fast to call
sequencetools.repeat_sequence_to_weight_exactly() on a list of
nonreduced fractions.

Optimization results: the change committed here makes duration interpretation
exponentially faster and shaves about 10% - 15% off of total interpretation
time of a nontrivial realworld specificationtools score example.