Optimized NonreducedFraction.__add__().
Local results: The method previously took 300 function calls.
The method now takes 35 function calls (when adding
to an integer or to a fraction with like denominator)
or 65 function calls (when adding to a fraction
with unlike denominator).

Realworld results: This translates to a change from 
9,733,944 function calls down to 8,118,160 function 
calls when interpreting a nontrivial segment of 
specificationtools score with a couple
dozen division select expressions. 
This equals a 17% reduction in function calls total.

It seems likely that all remaining arithmetic
operations implemented on NonreducedFraction
should be optimized in a similar way.