Optimized StartPositionedPayloadExpression.get_elements_that_satisfy_time_relation() again.
The method now caches results.

RESULTS:

    On specificationtools optimization target #5:
        before: 1,955,779 function calls
        after:  1,848,278 function calls

    On nontrivial cello music:
        before: 4,965,658 function calls
        after:  4,670,410 function calls

The performance improvements are modest only because the previous
optimization of the same code (using bisect) improved performance
to the degree that it did.

The optimization was straightforward and demonstrates that the caching
pattern can be easily integrated in the other places of the system
that will benefit from it.