Extended ScorePackageWrangler with new optimized method.
The new method is ScorePackageWrangler.list_visible_asset_importable_name_and_score_title_pairs().

The method builds up the main menu entries for the list of scores on the system in an efficient way.
because method calls get_tags() only once per score.

Repeated calls to get_tags() are slowing the system down throughout.

This new method will be hooked into the system in a future commit
and optimization results will be provided at that time.