Changed 'module_importable_name' globally to 'module_path. TERMINOLOGY update included.
NOTE: we will migrate package and module terminology to the following:

    Example PACKAGE PATHS:
        scoremanagertools.core.ScoreManagerObject                  
        scoremanagertools.core                                    
        scoremanagertools                                        

    Example PACKAGE NAMES:
        scoremanagertools
        core
        ScoreManagerObject

    Example MODULE PATH:
        scoremanagertools.core.ScoreManagerObject.ScoreManagerObject

    Example MODULE NAME:
        ScoreManagerObject

That is, we will work with the idea of a 'package path' that is analogous to a 'directory path'.
(Package paths MAY contain a dot similarly to how directory paths ALWAYS contain a slash.)

We will also work with the idea of a 'package name' that is analogous to a 'directory name'.
(Package names will NEVER contain a dot and directory names will NEVER contain a slash.)

We will also work with the idea of a 'module path' that is analagous to a 'file path'.
(Module paths MAY contain a dot similarly to how file paths ALWAYS contain a slash.)

Finally, we will work with the idea of a 'module name' that is analogous to a 'file name'.
(Module names will NEVER contain a dot and file names will NEVER contain a slash.)