Improved importtools.
In the past, importtools._import_public_names_from_path_into_namespace would
attempt to determine the module path of a given directory by locating the
last occurance of the ``root_package_name`` argument in that package's path
string.

Given a ``root_package_name`` of "abjad":

    /Users/josiah/Documents/Development/abjad/trunk/abjad/tools/notetools/Note
    =>
    abjad.tools.notetools.Note

...but:

    /Users/josiah/Documents/Development/abjad/trunk/abjad/tools/abjadbooktools/AbjadBookProcessor
    =>
    abjadbooktools.AbjadBookProcessor

Now, we search not simply for the ``root_package_name``, but that combined
with the os path separator character.