Introduced the concept of a 'filesystem path'. TERMINOLOGICAL note included.
MOTIVATION: the concepts of 'file path' and 'directory path' are clear.
So is the concept of 'package path'. (All three are types of path.)
But how do we refer collectively to 'file path' and 'directory path'
(while leaving out 'package path')? This is the motivation for introducing
the term 'filesystem path' as a term that generalizes both 'file path'
and 'directory path' (but which does not include 'package path').

NOTE: it follows that every asset has exactly one 'filesystem path'.
(And indeed this commit changes AssetProxy.asset_path to
AssetProxy.asset_filesystem_path.) It also follows that some assets
will also have a 'package path' while other assets will not
have a 'package path'.