Renamed experimental/specifiers directory. Terminological note included.
The new name is experimental/system_specifiers.

NOTE: this commit begins the process of clarifying what has up to now been
called 'score-internal' and 'score-external' assets.
The new naming regime will oppose 'system' resources to 'user' resources.
It remains to be seen if the score-internal and -external distinction
will persist or be supplanted by user / system.