Turned quick caching back on.
Py.test runtime decreases from 46 seconds to 37 seconds with caching turned on.

NOTE: caching will be left on during development (like now) and turned off for public releases.

This seems to represent the best comprise between the performance increases given by
caching and the potential maintenance overhead of working out all the cornercases of the cache.