Removed sketches from score manager.
Explanation: the motivation behind the idea of the sketch is to
allow the composer to quickly create a small but complete score:
a sketch. This means that sketches will wind up being basically
the same thing as scores. Which further means that sketches
can be safely ignored during the first several iterations of the
score manager: just create a score instead.

Sketches can be reintroduced to the score manager several point
versions later if the use case is still present.