Started TODO item of refactoring all user input-handling methods as dictionaries.
This commit refactors ScoreManager user-input handling as a dictionary.

EXPLANATION:

Here is a slightly edited version of ScoreManager._handle_main_menu_result() prior to this commit:

     def _handle_main_menu_result(self, result):
        if result == 'active':
            self._session.show_active_scores()
        elif result == 'all':
            self._session.show_all_scores()
        elif result == 'f':
            self.music_specifier_module_wrangler._run()
        elif result == 'fix':
            self.score_package_wrangler.fix_visible_assets()
        elif result == 'm':
            self.material_package_wrangler._run()
        elif result == 'mb':
            self._session.show_mothballed_scores()
        elif result == 'new':
            self.score_package_wrangler.make_asset_interactively()
        elif result == 'profile':
            self.score_package_wrangler.profile_visible_assets()
        elif result == 'py.test':
            self.run_py_test_on_all_user_scores()
        elif result == 'svn':
            self.manage_svn()

Here is the new implementation:

    def _handle_main_menu_result(self, result):
        if result in self.user_input_to_action:
            self.user_input_to_action[result](self)
        elif result in self.score_package_wrangler.list_visible_asset_packagesystem_paths():
            self.edit_score_interactively(result)

This compact implementation is possible because of the new ScoreManager user_input_to_action mapping:

    ### USER INPUT MAPPING ###

    user_input_to_action = {
        'active':   show_active_scores,
        'all':      show_all_scores,
        'f':        manage_specifiers,
        'fix':      fix_visible_scores,
        'm':        manage_materials,
        'mb':       show_mothballed_scores,
        'new':      make_new_score_interactively,
        'profile':  profile_visible_scores,
        'py.test':  run_py_test_on_all_user_scores,
        'svn':      manage_svn,
        }

ScoreManager.user_input_to_action is implemented as a class variable at *the very bottom*
of the ScoreManager classfile. This means that all of the ScoreManager actions will be defined
towards the top of the ScoreManger classfile and will therefore be in scope when
ScoreManage.user_input_to_action is defined.

OBSERVATION. This would appear to be a very clean way of modeling (and also declaring and encapsulating)
all of the UI functionality of the primary classes in the system.

PROSPECTUS. The idea going forward will be to refactor all UI-oriented classes this way.