Removed super(type(self), self).parent_class_method() pattern. INHERITANCE PATTERN DESCRIBED.
EXPLANATION. There is no generalized replacement. Instead you must use something like ...

    super(ListEditor, self).parent_class_method()

... with the current class (here ListEditor) written out in full.


Note first that this is a limitation of Python 2.x and that we will be able to say ...

    super().parent_class_method()

... in Python 3.x.


Note further that this is a well-discussed problem in the community.
The best summary of how to use super is Raymond Hettinger's blog post here:

    http://rhettinger.wordpress.com/2011/05/26/super-considered-super/

(And the comments in the comment section are just as useful
in understanding the value of super in 2.7 and 3.x as is the post itself.)


PROSPECTUS. In Python 2.x we have two choices for calling superclass methods.
To name the methods we will assume base class A:

    class A(object):
        def foo(self):
            ...

FIRST METHOD:

    class B(A):
        def foo(self):
            A.foo()

SECOND METHOD:

    class B(A): 
        def foo(self):
            super(B, self).foo()


Note first that the relevant typographic difference between the two methods
is that the first method requires the name of superclass A to be written out
in full while the second method requires the name of current class B to be written out
in full.

Note next that neither method affords a way to refer indirectly to both superclass A
*and* current class B. (NOTE that this is the real goal: a way of refering to superclass
methods that involves refering to neither the superclass nor the current class explicitly.
This is totally doable in cases of single inheritance. And, in fact, super() affords this
behavior in Python 3.x. But in Python 2.x we are forced to choose between naming
either the superclass or the current class. The discussion below examines the relevants
merits and annoyances of each of these two methods.)

For most of the prior years of Abjad development we've used the first method.
But going forward the convention will be to use the SECOND METHOD. Why?
Because the second method provides a sliver more indirection than the first method.
Raymond makes this point in the comments section of the blog post linked to above:
the second method allows the bases of current class B to be swapped out
without causing any problems to B. This is a concrete benefit not available
under the first method.

Also, making pervasive use of super() brings us closer to how things will
look when we finally migrate to Python 3. While not a direct benefit to the
current implementation, the combination of a slight benefit of indirection
together with the eventual benefits after migration make the second
pattern now prefered.