Improved timesignaturetools.establish_metrical_hierarchy() docstring.
When writing documentation where the output of the same function is compared with slight variations in keyword arguments, it is always best to put every argument to that function on its own line.  This makes it considerably easier for those reading the documentation to quickly determine what has changed from one example to the next.