Extended menu classes with temporary is_modern=False keyword.
The keyword will remain in the system during migration from
this type of menu building ...

    menu_section.append(('add', 'svn add scores'))
    menu_section.append(('ci', 'svn commit scores'))
    menu_section.append(('st', 'svn status scores'))
    menu_section.append(('up', 'svn update scores'))

... to this type of menu building:

    menu_section.append(('svn add scores', 'add'))
    menu_section.append(('svn commit scores', 'ci'))
    menu_section.append(('svn status scores', 'st'))
    menu_section.append(('svn update scores', 'up'))

The older form is (key, display_string).

The modern form is (display_string, key).

The older form is deprecated.

The reason for the change is that all menu entries comprise
a display string while only some menu entires comprise a key.
This means that the display string is the fundamental input parameter
necessary to create a menu entry and should always be the first
parameter.