Modernized score manager main menu.
NOTE: you must trash your ~/score_manager/cache.py file
for tests to work correctly after this update.

(Why? Because the contents of the ~/score_manager/cache.py file
is a persistent copy of the tokens that make up the score manager
main menu. The format of those tokens has changed in this commit.
This invalidates the cache currently resident on your disk.)