Enabled indent-based folding in etc/editors/vimrc file.
IDE NOTE: note that you can now press ...

    <leader>zm

... to fold up basically all the methods in a classfile.
(The one caveat here is that the mapping doesn't yet work
for stuff headed by decorators.)

Then press ...

    zR

... when you want to unfold the entire file (and therefore
see everything again).

Note that the first command involves the leader key and that the second
command doesn't. The reason for this is that <leader>zm is a mapping
added directly to our etc/editors/vimrc file while zR is native vim.

Also note that the fundamental setting that makes this stuff work
is the "set foldmethod=indent" line in etc/editors/vimrc.
This tells vim to fold lines that have the same indent level.
Which is basically perfect for Python.

Chapter 28 of the vim reference manual talks about folding.
There's quite a lot more there that we can probably use
very effectivel.

ALSO NOTE: that once you have called <leader>zm in a file,
you can toggle between the folded and unfolded view of the file
with zi (which is also built-in vim).