Added foldlevel=1 setting to etc/editors/vimrc file.
IDE: This causes files to open with methods folded.

This looks to be considerably more useful than opening
files in the normal way (with everything unfolded) because
the structure of the file is more or less visible immediately.

NOTE: that you can use "zi" to toggle between folded and unfolded views.