Added foldminlines=0 setting to etc/editors/vimrc file.
IDE: vim's default value for this setting is foldminlines=1.
This means that vim by default doesn't fold methods that
contain only a 1-line implementation. By changing
foldminlines to 0 we tell vim that we want vim to fold
all methods (even those with 1-line implementations).

Note that the comment in a previous commit about folding
not yet working correctly with decorators with unfounded:
what was going on is that vim was not folding a number of
decorated methods with 1-line implementations. As of this
commit vim now folds all methods; the decorator status
of the method never played a role in what was observed.