Added new autocommand to etc/editors/virmc file.
The new autocommand is this:

  au BufWinEnter *.py silent loadview                                             

The command executes vim's ":loadview" function
every time we open a new *.py file in vim.

What does vim's ":loadview" function do?
The function looks on the filesystem for any so-called
'view' file you earlier created using vim's ":makeview" function.

You can read about :makeview in vim's online help.
But the reason :makeview is important right now is that
:makeview causes vim to store folds that you have manually
created for a file. This is can add the finishing touches on the way
that you want a file to fold up for an easy overview.
More specifically, we now have foldmethod=indent set in the 
etc/editors/vimrc file. What you can then do is open the classfile
that you're currently spending time in and run ":set foldmethod=manual".
This will not remove the folds automatically created earlier
by foldmethod=indent (when you first opened the file). Once
you ":set foldmethod=manual" you can highlight a chunk of text
(in visual mode, for example) and then type "zf" to fold that chunk
of text. (This is good to do with all import statements at the top
of a file, for example. And also for the main docstring to a class.)
Finally, after you've made a couple of manual folds (using "zf")
you can run ":makeview" to cause vim to write out a (persistent)
'view' file to disk for the classfile you've just been editing.
The next time you open that class file, the autocommand included
with this commit will cause vim to read the 'view' file and reestablish
both the folds created automatically by foldmethod=indent and also
the folds you manually created with foldmethod=manual.