Removed problematic paste setting from etc/editors/vimrc file.
NOTE: this fixes the paste-says-its-on-but-isn't-actually-working problem.

The problematic line was here:

    filetype plugin indent on

The problematic line was doing two things:

    1. making plugins load automatically (based on file type)
    2. reading indent files automatically (based on file type)

It turns out that loading plugins automatically is OK
but that reading indent files automatically was causing our paste problem.

Here's the new setting:

    filetype plugin on

Pasting from terminal into vi should work now (when in paste mode).