Improved viewcode annotations, and turned viewcode back on for the mainline docs.
See durationtools.Duration's methods documentation for a good example of the
extension in action.

We may need to extract the annotating logic from the viewcode extension, in
order to speed up doc rendering, or find a means to prevent viewcode from
writing out HTML files if it can be avoided.