DEVELOPER FUNCTIONALITY: codebase is now tagged for navigation in Vim.
This commit checks in three things:

    1. a new script that tags the entire codebase
    2. the tags file created by the script
    3. an update to the etc/editors/vimrc that tells Vim about the tags file

The new functionality is basically a miracle:

    * put your cursor on any class name and press <C-]> to be
      instantly transported to the definition of the class
      in its remote classfile

    * press <C-t> to return to wherever you were before

Note that <C-]> and <C-t> are the two-key chords used to navigate
the links in the Vim documentation, which should make the navigation
feel familiar.

More information is available at :help tags. But it's enough to
simply start pressing <C-]> and <C-t> and be amazed.

(Note, too, that the tags file is in $ABJADROOT/tags. The file is cleartext
and follows an obvious format for storing the patterns Vim uses as tags.
Also note that the filesystem paths stored in the file are all relative
to $ABJADROOT. This means that the tags file needs to stay in $ABJADROOT
and should not be moved elsewhere, such as to etc/editors, for example.)