Extended instruments in instrumenttools with 'default_performer_abbreviation'.
Consider the clarinet in B-flat and also the clarinet in A. 
The abbreviations for these two instrument names should be
'cl. in B-flat' and 'cl. in A', or something comparable.
But the abbreviation for the *player* of these two instruments
should be only 'cl.' (or comparable).

This commit introduces the default_performer_abbreviation property
to model this distinction.

This changes the output of scoretools.list_primary_performer_names().
But doesn't need to be mentioned in the 2.13 change log.