IDE: vim-pythontextobj plugin now recommended.
Explanation: the plugin is available at ...

    https://github.com/natw/vim-pythontextobj.git

... and was authored by Nat Williams.

The plugin is a lightweight 135-line Vim-script file.

The plugin provides two pairs of Vim text objects:

    * af, if
    * ac, ic

Explanation:

    * af selects "a function (with following blank lines)"
    * if selects "inner function (without following blank lines)"

    * ac selects "a class (with following blank lines)"
    * ic selects "inner class (without following blank lines)"

This means that you can, for example, use ...

    vaf

... to visually select all of a class method in precisely
the same way that you can use "vap" to select all of a paragraph.

Note, too, that the plugin also works with decorators!

No further configuration appears to be necessary in our etc/editors/vimrc file.

All you have to do is change to ~/.vim/bundle and clone the repository.