IDE: adjusted arrow key remapping in etd/editors/vimrc file.
The arrow keys are now disabled in all of Vim's modes that allow for mapping.

This includes insert mode.

However, note that there's an important caveat that Vim insists that all
insert mode mappings are disabled when paste mode is turned on.
This means that arrow keys will always work when paste mode is turned on.

Accordingly, this commit sets nopaste. This is different than
what we're used to, but is probably better: the Vim community 
seems predominantly to leave paste mode turned off and only to toggle
it on immediately before a paste operation.

So let's run with nopaste for a while and see how it goes.