Updated abjad-grep script.
Changed the core command from ...

    ack --ignore-dir=docs pattern *

... to ...

    ack --ignore-dir=docs pattern

instead.

The second version omits the star. This is important.
What goes on if you pass a star to ack is that the
star is expanded by the shell to all files in the current directory.
If the current directory contains .pyc or other crufty files
then ack will search all of those files. And this is true even
if you pass ack an exclusion pattern specifically telling it NOT
to search files of that type!

CONCLUSION: the conclusion is that you don't need to pass a * into
ack the way that you do into a recursive call to grep.