Moved tonalitytools functionality.
Old function:

    * tonalitytools.make_all_notes_in_ascending_and_descending_diatonic_scale()

New method:

    * tonalitytools.Scale.make_score()

Doesn't need to be noted in the 2.13 change log.

The tonalitytools functions are now rather more trim than before:

    tonalitytools.analyze_chord
    tonalitytools.analyze_incomplete_chord
    tonalitytools.analyze_incomplete_tonal_function
    tonalitytools.analyze_tonal_function
    tonalitytools.are_scalar_notes
    tonalitytools.are_stepwise_ascending_notes
    tonalitytools.are_stepwise_descending_notes
    tonalitytools.are_stepwise_notes
    tonalitytools.is_neighbor_note
    tonalitytools.is_passing_tone
    tonalitytools.is_unlikely_melodic_diatonic_interval_in_chorale

MODEL: it seems like the is_* and are_* predicates listed above should
be bound to something. Probably some type of selection. But adding
to the Selection namespace seems inappropriate for this sort of decidedly
second-order package.

MODEL: does it make sense to implement a TonalAnalysisSelection?
The class could be returned by a new tonalitytools.select() method
designed to select an arbitrary part of a score and then provide
access to the functions in the 'analytic interface' listed above.