Updated ClassDocumenter banner.
ClassDocumenter now outputs 'Attribute summary' instead of 'Attributes
summary'.

Where possible, it's nice to prefer noun phrases built of only
uninflected components for all components except the rightmost one in the
chain. This is the basis for prefering 'attribute' to 'attributes':
'attribute' is the uninflected ("dictionary form" or "citation form")
of the word. This uninflected-except-for-last rule is certainly not 
hard and fast; but where we can convincingly use it, it is probably
a clean option.

Docs have not been rebuilt in this commit;
docs will be rebuilt in a coming commit.