Renamed MultiMeasureRest. The new name is MultimeasureRest.
The name MultimeasureRest differs from the LilyPond MultiMeasureRest grob.

But we have to diverge from LilyPond in this case because
treating bound mophemes (in this case the multi- prefix) as
independent words just isn't tenable.

(Note that our spelling follows LilyPond spelling in cases 
of compounding choice: we spell "note head" and "NoteHead", for example,
instead of "notehead" or "note-head". English provides no
process for determining the "spelling" of compounds, so we follow
the choices of the LilyPond developers in these cases.)

NOTE: this is a two-part commit. This commit temporarily renames
MultiMeasureRest to TemporaryRest. A subsequent commit will
rename TemporaryRest to MultimeasureRest. This commit will
not interpret and should not be used to compile score.