Removed all __cmp__ test files.
This is a preliminary commit in preparation for removing the
definitions of __gt__, __ge__, __lt__ and __le__ from AbjadObject.

The __cmp__ test files removed in this commit all looked like this:

    def test_Tuplet___cmp___01():

        tuplet = Tuplet(Fraction(2, 3), "c'8 d'8 e'8")

        assert tuplet == tuplet
        assert not tuplet != tuplet

        comparison_string = 'tuplet <  tuplet'
        assert py.test.raises(NotImplementedError, comparison_string)
        comparison_string = 'tuplet <= tuplet'
        assert py.test.raises(NotImplementedError, comparison_string)
        comparison_string = 'tuplet >  tuplet'
        assert py.test.raises(NotImplementedError, comparison_string)
        comparison_string = 'tuplet >= tuplet'
        assert py.test.raises(NotImplementedError, comparison_string)
        
There were about 40 such tests for about a dozen classes,
all of the same structure.

When we move to Python 3.0 we'll get the refuse-to-compare behavior
tested for here. In the meantime we'll open up large amounts of 
space in the docs if we get rid of the four comparison operations.

This commit deprecates (but doesn't remove) __gt__, __ge__,
__lt__ and __le__ on AbjadObject.