Added MinimalSelection class to selectiontools package.
Rewired all selection classes to inherit from MinimalSelection.

REFACTORING: MinimalSelection is introduced in this commit as
a type of newly created base class for all selection classes.
MinimalSelection will eventually be renamed to Selection;
the current Selection class will eventually be renamed to something
like ComponentSelection or ScoreMutationSelection.

REFACTORING: the point in doing all of this is to remove functionality
from certain selection classes. There is no reason, for example,
for the "vertical" selection classes (Lineage, Parentage, Descendants)
to implement duration attributes like duration and timespan (and the current
commit in fact moves duration attributes from Selection to
HorizontalSelection for this reason). Likewise, there is no reason
for TonalAnalysisSelection to expose attach_marks(), detach_marks(),
get(), get_spanners() and so on: when the composer makes a tonal
analysis selection, all that should appear in the interface of the
resulting TonalAnalysisSelection is a highly targeted list of methods
useful for doing tonal analysis. This is now the case as of this commit:

    TonalAnalysisSelection.analyze_chords()
    TonalAnalysisSelection.analyze_incomplete_chords()
    TonalAnalysisSelection.analyze_incomplete_tonal_functions(key_signature)
    TonalAnalysisSelection.analyze_neighbor_notes()
    TonalAnalysisSelection.analyze_passing_tones()
    TonalAnalysisSelection.analyze_tonal_functions(key_signature)
    TonalAnalysisSelection.are_scalar_notes()
    TonalAnalysisSelection.are_stepwise_ascending_notes()
    TonalAnalysisSelection.are_stepwise_descending_notes()
    TonalAnalysisSelection.are_stepwise_notes()

This was made possible in a single step by rewiring TonalAnalysisSelection
to inherit from MinimalSelection instead of from (the much bulkier)
Selection.