Removed 'preprolated_duration' property from all components.
EXPLANATION: composers use 'duration' (and, perhaps, 'duration_in_seconds')
almost exclusively in composition. The concept of preprolated duration is
crucial to the *implementation* of classes and methods everywhere in the
system. But it is now clear that composers can be shielded from the
concept in the public API.

Subsequent commits will remove 'preprolated_duration' from the public
interface of spanner classes and selection classes.

Developers may still access the preprolated durations of all objects
systemwide with the private '_preprolated_duration' properties.

CHANGE: should be noted in the 2.13 change log.