Removed inspect.py module from selectiontools package.
The module aliased Python's built-in 'inspect' module
when, for example, "ajv dev" or py.est were run from within the
abjad/tools/selectiontools directory.

(Why? Because the line "import inspect" would then resolve
to selectiontools/inspect.py rather than to the built-in inspect
module.)

Reimplemented the functionality as a static method on Inspector.

    Inspector.inspect()

This avoids a custom module aliasing a built-in module because there
is no prohibition against a custom *method* being named the same
as a built-in module.

All tests pass. But this still feels
like a suboptimal solution because selectiontools.select()
and selectiontools.Inspector.inspect() are accessed in different ways.

Too bad built-in 'inspect' is squatting on valuable lexical territory.

Even worse that neither Python 2 nor Python 3 offers a satisfactory
way to demand import from the built-in namespace (as opposed to starting
import search with the current directory).

Why doesn't sys.modules preserve superpackage names for
namespace disambiguation?