Harmonizing container insert methods.
The following three methods and functions all currently exist:

    1.  Container.__setitem__(i, expr)
    2.  Container.insert(i, component)
    3.  containertools.insert_component(
            container, i, component, fracture_spanners=False)

The three methods relate each other in the following ways:

The Container.__setitem__() logic is the most fundamental. We see that
this is the case because #2 and #3 are both implemented at least partially
in terms of #1.

#3 is a true generalization of #2. What this means is that #3 provides
exactly the same functionality #2 plus an extra chunk of entirely new
functionality. Specifically, calling #2 is now implemented (as of this
commit) as an exact call to #3 with fracture_spanners=False. This means
that #2 and #3 are redundant with respect to each other. The refactoring
strategy here will be to remove #3 and extend #2 with a fracture_spanners=False
keyword.

The following will remain:

    1. Container.__setitem__(i, expr)
    2. Container.insert(i, component, fracture_spanners=False)

It's actually kind of a shame that __setitem__ refuses keywords because
a __setitem__(i, expr, fracture_spanners=False) would fit out uses
perfectly. But the core Python syntax doesn't allow for the keyword
extension of special methods.