Added ContiguouSelection to selectiontools.
ContiguousSelection is currently a stub.

The class models time-contiguous selections of components;
SliceSelection is now just a special type of ContiguousSelection.

ContiguousSelection will act as the base class for the following
five types of selection:

    1. time-contiguous selections of leaves
    2. time-contiguous selections of tuplets
    3. time-contiguous selections of measures
    4. time-contiguous selections of mixed classes of components
    5. same as 4, with all in the same parent (ie, SliceSelection)

NEW: this class and all others in selectiontools should be explained
in the 2.13 release notes.