Extended selectiontools.select_leaves() and Container.select_leaves().
First, Container.select_leaves() is now implemented as an exact
wrapper around selectiontools.select_leaves().

Second, both methods now implement an allow_discontiguous_leaves=False keyword.
The keyword is set to false by default; this means that calling 
foo.select_leaves() is expected to return a time-contiguous selection
of leaves in the form of a ContiguousLeafSelection and that when leaves
are not contiguous an exception will be raised.

On the other hand, when allow_discontiguous_leaves=True then discontiguous
leaves may be returned in a FreeLeafSelection and no exception will
be raised.

MODEL: this basically forces a type of very deep disambiguation that's
been many years in the making. Calling foo.select_leaves() has historically
been a blind wrapper around iterationtools.iterate_leaves_in_expr(). This
is fine except that iterationtools.iterate_leaves_in_expr() will *wrap around*
staff and other context boundaries. Why? Because the iteration function
is simply a depth-first traversal of the score tree. But much usage
of foo.select_leaves() looks like a type of *thread selection*, e.g.
in calls like ...

    spannertools.BeamSpanner(foo.select_leaves())

... which very clearly assume that the output of foo.select_leaves()
will be nicely contiguous.

As of this commit all calls to (unqualified) foo.select_leaves()
*do* return nicely contiguous output. Older use of foo.select_leaves()
to blindly iterate foo in search of leaves can still be made
with foo.select_leaves(allow_discontiguous_leaves=True).