Tweaked ContainmentSignature comparison logic.
Previously: containment signatures compared false when roots compared false.
Also, parentage signatures compared false when roots compared false.

Now: containment signatures no longer compare roots.
On the other hand, parentage signatures *do* continue to compare roots.

Explanation: the components ...

    [Note("c'4"), Measure((2, 8), "d'8 e'8")]

... currently do not pass ...

    componenttools.all_are_components_in_same_thread()

... which is a problem.

This commit and the ones after it help fix that problem.

The current commit establishes that an orphan note and orphan measure
are contained in the same way. This means that an orphan note and an
orphan measure can be spanned.