Rewrote thread-checking function.
    * componenttools.all_are_thread_contiguous_components()

The function was way to difficult to understand.

This commit changes a small number of core things in order to rewrite
the function.

The behavior of a corner case is also changed with this commit.

Consider this score:

    \new Staff {
        c'8
        d'8
        {
        }
        e'8
        f'8
    }

Question: should the four notes in the staff be consider thread-contiguous?

The four notes were previously considered thread-contiguous;
the four notes are no longer considered thread-contiguous.

Consider this second score:

    \new Staff {
        c'8
        d'8
        \new Voice {
            c''8
            d''8
        }
        e'8
        f'8
    }

The four notes in the fourth octave were previously considered thread-
contiguous; the four notes in the fourth octave are no longer considered
thread-contiguous.

Doesn't need to be included in the change log.