Added logical voice component getter.
BACKGROUND: What is a namesake?

The answer isn't clear. And it would be happy if we could remove the
concept of the namesake entirely in favor of the concept of logical voice.

It looks like the namesakes of note_1 are all the other notes in 
the same logical voice as note_1. Likewise, all the namesakes of 
rest_1 are all the other rests in the same logical voice as rest_1.

(The one place where the concept of the namesake may differ from the
concept of the logical voice is that contexts like staves can apparently
have namesakes; by way of comparison, note that contexts like staves
are not included in any logical voice.)

The idea of the namesake appears in only a few places in the system.
The two most important are these:

    * iterationtools.iterate_namesakes_from_component()
    * Component._get_namesake()

Of these, Component._get_namesake() is implemented as only
a wrapper around iterationtools.iterate_namesakes_from_component().

A handful of other functions made calls to Component._get_namesake():

    * containertools.fuse_like_named_contiguous_containers()

    * measuretools.get_next_measure_from_component()
    * measuretools.get_previous_measure_from_component()
    * VerticalMoment.next_vertical_moment
    * VerticalMoment.previous_vertical_moment
    * TonalAnalysisSelection._is_neighbor_note()
    * TonalAnalysisSelection._is_passing_tone()

This commit implements Component._get_in_my_logical_voice() as a wrapper
around iterationtools.iterate_logical_voice_from_component()
instead of iterationtools.iterate_namesakes_from_component().

This commit also replaces all calls to Component._get_namesake()
with calls to Component._get_in_my_logical_voice in these functions:

    * measuretools.get_previous_measure_from_component()
    * VerticalMoment.next_vertical_moment
    * VerticalMoment.previous_vertical_moment
    * TonalAnalysisSelection._is_neighbor_note()
    * TonalAnalysisSelection._is_passing_tone()

This shows that iterationtools.iterate_logical_voice_from_component() is
equivalent ot iterationtools.iterate_namesakes_from_component() for these
methods.

What remains is ...

    * containertools.fuse_like_named_contiguous_containers()

... which fails with Component._get_in_my_logical_voice() but passes
with Component._get_namesake().

Future commits will do more work to deprecate namesake in favor of logical
voice.